Moving column-select from uniprotkb to shared.
