+++ src/app/state/rootReducer.ts
@@ -2,12 +2,10 @@ import { combineReducers, Reducer } from 'redux';
 
 import { RootState, RootAction } from './rootInitialState';
 
-import resultsReducers from '../../uniprotkb/state/resultsReducers';
 import messagesReducers from '../../messages/state/messagesReducers';
 import toolsReducers from '../../tools/state/toolsReducers';
 
 const appReducer = combineReducers<RootState>({
-  results: resultsReducers,
   messages: messagesReducers,
   tools: toolsReducers,
 });
