+++ src/shared/components/column-select/ColumnSelectDragDrop.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import { DragDropContext, Droppable, Draggable } from 'react-beautiful-dnd';
 import { CloseIcon } from 'franklin-sites';
-import { getBEMClassName as bem } from '../../../shared/utils/utils';
-import { SelectedColumn } from '../../types/resultsTypes';
-import { UniProtKBColumn } from '../../types/columnTypes';
+import { getBEMClassName as bem } from '../../utils/utils';
+import { SelectedColumn } from '../../../uniprotkb/types/resultsTypes';
+import { UniProtKBColumn } from '../../../uniprotkb/types/columnTypes';
 import './styles/column-select-drag-drop.scss';
 
 type ColumnSelectDragDropProps = {
