+++ src/shared/components/column-select/__tests__/__mocks__/structuredResultFieldsData.json
