+++ src/shared/components/column-select/styles/column-select-drag-drop.scss
