+++ src/uniprotkb/components/customise-table/__tests__/CustomiseTableContainer.spec.tsx
@@ -3,10 +3,7 @@ import { fireEvent } from '@testing-library/react';
 import CustomiseTableContainer from '../CustomiseTableContainer';
 import initialState from '../../../../app/state/rootInitialState';
 import renderWithRedux from '../../../../shared/__test-helpers__/RenderWithRedux';
-import * as resultsActions from '../../../state/resultsActions';
 import '../../__mocks__/mockApi';
-
-const updateTableColumns = jest.spyOn(resultsActions, 'updateTableColumns');
 
 describe('CustomiseTableContainer component', () => {
   let renderedWithRedux, goBack;
