+++ src/uniprotkb/components/download/__tests__/DownloadView.spec.tsx
@@ -6,7 +6,7 @@ import { FileFormat } from '../../../types/resultsTypes';
 import renderWithRedux from '../../../../shared/__test-helpers__/RenderWithRedux';
 import initialState from '../../../../app/state/rootInitialState';
 
-// FIXME: Doing network call through ColumnSelectContainer's logic and Redux
+// FIXME: Doing network call through ColumnSelect's logic and Redux
 // FIXME: logic. Might need to mock after removing Redux logic to load data.
 // NOTE: Causing error when no network available, when any 2 tests below run 🤷🏽‍♂️
 
@@ -70,7 +70,7 @@ describe('DownloadView component', () => {
   });
 
   test('should call onFileFormatChange when file format select is changed', () => {
-    const { getByTestId, queryByText } = renderedWithRedux;
+    const { getByTestId } = renderedWithRedux;
     const formatSelect = getByTestId('file-format-select');
     fireEvent.change(formatSelect, { target: { value: FileFormat.gff } });
     expect(props.onFileFormatChange).toHaveBeenCalled();
